@@ -1,7 +1,7 @@
 package com.example.workerserver.infrastructure.entity
 
 import java.time.ZonedDateTime
-import javax.persistence.*
+import jakarta.persistence.*
 
 @Entity
 @Table(name = "BulkRequestLog")
