@@ -17,6 +17,6 @@ class WebSocketHandler: WebSocketMessageBrokerConfigurer {
     }
 
     override fun registerStompEndpoints(registry: StompEndpointRegistry) {
-        registry.addEndpoint("/ws").withSockJS()
+        registry.addEndpoint("/bulk").withSockJS()
     }
 }
