@@ -1,12 +1,7 @@
 package com.example.workerserver.infrastructure.entity
 
+import jakarta.persistence.*
 import java.time.ZonedDateTime
-import javax.persistence.Column
-import javax.persistence.Entity
-import javax.persistence.GeneratedValue
-import javax.persistence.GenerationType
-import javax.persistence.Id
-import javax.persistence.Table
 
 @Entity
 @Table(name = "BulkProcessLog")
